requirements.txt
azure-search-documents==11.4.0